test(voice-call): isolate caller-turn regression
